Changed page to be rendered on the server
